auto_arbitrage_bot/requirements.txt
ccxt>=4.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
websockets>=11.0.0
requests>=2.28.0
numpy>=1.24.0
pandas>=1.5.0